--- a/src/compiler/scala/tools/nsc/typechecker/Contexts.scala
+++ b/src/compiler/scala/tools/nsc/typechecker/Contexts.scala
@@ -129,7 +129,7 @@ trait Contexts { self: Analyzer =>
         case s              =>
           getModuleIfDefined(s) orElse
           getPackageObjectIfDefined(s) orElse
-          getPackageIfDefined(s) orElse {
+          getPackageIfDefined(TermName(s)) orElse {
             error(s"bad preamble import $s")
             NoSymbol
           }