--- a/src/reflect/scala/reflect/api/Names.scala
+++ b/src/reflect/scala/reflect/api/Names.scala
@@ -41,19 +41,19 @@ import scala.language.implicitConversions
  *  @group ReflectionAPI
  */
 trait Names {
-  /** An implicit conversion from String to TermName.
-   * Enables an alternative notation `"map": TermName` as opposed to `TermName("map")`.
-   * @group Names
-   */
-  @deprecated("use explicit `TermName(s)` instead", "2.11.0")
-  implicit def stringToTermName(s: String): TermName = TermName(s)
-
-  /** An implicit conversion from String to TypeName.
-   * Enables an alternative notation `"List": TypeName` as opposed to `TypeName("List")`.
-   * @group Names
-   */
-  @deprecated("use explicit `TypeName(s)` instead", "2.11.0")
-  implicit def stringToTypeName(s: String): TypeName = TypeName(s)
+//  /** An implicit conversion from String to TermName.
+//   * Enables an alternative notation `"map": TermName` as opposed to `TermName("map")`.
+//   * @group Names
+//   */
+//  @deprecated("use explicit `TermName(s)` instead", "2.11.0")
+//  implicit def stringToTermName(s: String): TermName = TermName(s)
+//
+//  /** An implicit conversion from String to TypeName.
+//   * Enables an alternative notation `"List": TypeName` as opposed to `TypeName("List")`.
+//   * @group Names
+//   */
+//  @deprecated("use explicit `TypeName(s)` instead", "2.11.0")
+//  implicit def stringToTypeName(s: String): TypeName = TypeName(s)
 
   /** The abstract type of names.
    *  @group Names