--- a/src/reflect/scala/reflect/api/Names.scala
+++ b/src/reflect/scala/reflect/api/Names.scala
@@ -55,30 +55,82 @@ trait Names {
 //  @deprecated("use explicit `TypeName(s)` instead", "2.11.0")
 //  implicit def stringToTypeName(s: String): TypeName = TypeName(s)
 
+  type NameTable <: NameTableApi
+
+  val nameTable: NameTable
+
   /** The abstract type of names.
    *  @group Names
    */
-  type Name >: Null <: AnyRef with NameApi
+  type Name >: Null <: AnyRef with nameTable.NameApi
 
   /** The abstract type of names representing terms.
    *  @group Names
    */
-  type TypeName >: Null <: TypeNameApi with Name
+  type TypeName >: Null <: nameTable.TypeNameApi with Name
+  /** The abstract type of names representing types.
+    *  @group Names
+    */
+  type TermName >: Null <: nameTable.TermNameApi with Name
 
-  /** Has no special methods. Is here to provides erased identity for `TypeName`.
-   *  @group API
-   */
-  trait TypeNameApi
+  /** The API of Name instances.
+    *  @group API
+    */
+  type NameApi <: nameTable.NameApi
+
+  /** Create a new term name.
+    *  @group Names
+    */
+  @deprecated("use TermName instead", "2.11.0")
+  def newTermName(s: String): TermName
+
+  /** Creates a new type name.
+    *  @group Names
+    */
+  @deprecated("use TypeName instead", "2.11.0")
+  def newTypeName(s: String): TypeName
+
+  /** The constructor/extractor for `TermName` instances.
+    *  @group Extractors
+    */
+  val TermName: TermNameExtractor
+
+  /** An extractor class to create and pattern match with syntax `TermName(s)`.
+    *  @group Extractors
+    */
+  abstract class TermNameExtractor {
+    def apply(s: String): TermName
+    def unapply(name: TermName): Option[String]
+  }
+
+  /** The constructor/extractor for `TypeName` instances.
+    *  @group Extractors
+    */
+  val TypeName: TypeNameExtractor
+
+  /** An extractor class to create and pattern match with syntax `TypeName(s)`.
+    *  @group Extractors
+    */
+  abstract class TypeNameExtractor {
+    def apply(s: String): TypeName
+    def unapply(name: TypeName): Option[String]
+  }
+}
+
+abstract class NameTableApi {
+  /** The abstract type of names.
+    *  @group Names
+    */
+ type Name >: Null <: AnyRef with NameApi
+
+ type TypeName >: Null <: TypeNameApi with Name
+
 
   /** The abstract type of names representing types.
    *  @group Names
    */
   type TermName >: Null <: TermNameApi with Name
 
-  /** Has no special methods. Is here to provides erased identity for `TermName`.
-   *  @group API
-   */
-  trait TermNameApi
 
   /** The API of Name instances.
    *  @group API
@@ -116,6 +168,15 @@ trait Names {
      */
     def encodedName: Name
   }
+  /** Has no special methods. Is here to provides erased identity for `TypeName`.
+    *  @group API
+    */
+  trait TypeNameApi extends NameApi
+  /** Has no special methods. Is here to provides erased identity for `TermName`.
+    *  @group API
+    */
+  trait TermNameApi extends NameApi
+
 
   /** Create a new term name.
    *  @group Names
@@ -129,29 +190,4 @@ trait Names {
   @deprecated("use TypeName instead", "2.11.0")
   def newTypeName(s: String): TypeName
 
-  /** The constructor/extractor for `TermName` instances.
-   *  @group Extractors
-   */
-  val TermName: TermNameExtractor
-
-  /** An extractor class to create and pattern match with syntax `TermName(s)`.
-   *  @group Extractors
-   */
-  abstract class TermNameExtractor {
-    def apply(s: String): TermName
-    def unapply(name: TermName): Option[String]
-  }
-
-  /** The constructor/extractor for `TypeName` instances.
-   *  @group Extractors
-   */
-  val TypeName: TypeNameExtractor
-
-  /** An extractor class to create and pattern match with syntax `TypeName(s)`.
-   *  @group Extractors
-   */
-  abstract class TypeNameExtractor {
-    def apply(s: String): TypeName
-    def unapply(name: TypeName): Option[String]
-  }
 }