--- a/test/benchmarks/src/main/scala/reflect/NamesBenchmark.scala
+++ b/test/benchmarks/src/main/scala/reflect/NamesBenchmark.scala
@@ -16,7 +16,6 @@ import java.util.concurrent.TimeUnit
 
 import org.openjdk.jmh.annotations._
 
-import scala.collection.Outer.Inner
 import scala.reflect.internal.Names
 
 @BenchmarkMode(Array(Mode.AverageTime))