--- a/test/files/run/reflection-names.check
+++ b/test/files/run/reflection-names.check
@@ -1,4 +1,4 @@
 (java.lang.String,bc)
-(scala.reflect.internal.Names$TermName_R,bc)
-(scala.reflect.internal.Names$TypeName_R,bc)
-(scala.reflect.internal.Names$TypeName_R,bc)
+(scala.reflect.internal.NameTable$TermNameImpl,bc)
+(scala.reflect.internal.NameTable$TypeNameImpl,bc)
+(scala.reflect.internal.NameTable$TypeNameImpl,bc)