… test/benchmarks/src/main/scala/reflect/NamesBenchmarkInsertNew.scala
… test/junit/scala/reflect/internal/NamesTest.scala
… test/junit/scala/reflect/internal/TypesTest.scala
… test/junit/scala/tools/nsc/symtab/StdNamesTest.scala