inherently threadsafe, and could be shared between Runs - but not in this PR

the main motivation is that in our internal compilation we see that the name table is overloaded and has chains up to 500 long, so looked to see if we could make it faster

This is very much a first cut, but it seems to be faster for the simple JMH test attached
NamesBenchmark - looks up names that exist in the table
NamesBenchmarkInsertNew - inserts names from strings ( that have to calculate their hashcode)

```
OLD
[info] Benchmark                           Mode  Cnt       Score      Error  Units
[info] NamesBenchmark.createBoth           avgt   20   82178.994 ▒ 2159.234  ns/op
[info] NamesBenchmark.createTerm           avgt   20   23055.097 ▒ 1140.830  ns/op
[info] NamesBenchmark.createType           avgt   20   35959.945 ▒ 2379.958  ns/op
[info] NamesBenchmarkInsertNew.createBoth  avgt   20  425193.881 ▒ 2876.201  ns/op
[info] NamesBenchmarkInsertNew.createTerm  avgt   20  286452.250 ▒ 3880.833  ns/op
[info] NamesBenchmarkInsertNew.createType  avgt   20  416672.282 ▒ 3482.710  ns/op

NEW 
[info] Benchmark                           Mode  Cnt       Score      Error  Units
[info] NamesBenchmark.createBoth           avgt   20   25057.350 ▒  136.583  ns/op
[info] NamesBenchmark.createTerm           avgt   20   20373.252 ▒ 1058.120  ns/op
[info] NamesBenchmark.createType           avgt   20   20881.439 ▒  453.519  ns/op
[info] NamesBenchmarkInsertNew.createBoth  avgt   20  139702.194 ▒ 3616.623  ns/op
[info] NamesBenchmarkInsertNew.createTerm  avgt   20  134879.891 ▒ 2139.333  ns/op
[info] NamesBenchmarkInsertNew.createType  avgt   20  137091.462 ▒ 5214.178  ns/op
```

I did see some much faster times when using a class instead of a trait , but could not see a way to make that work without changing a lot

I removed some deprecated APIs and proposed some more that could be inlined in the code IMO, but left these so as to ease review

I have not looked at PDO yet, so there is probably more to gain still