@retronym 2019-04-10
The build is failing with an OOME when using the new compiler to compile the library. That's happening super early in the initalization of `Run`:

```
error: java.lang.OutOfMemoryError: Java heap space
	at scala.reflect.ManifestFactory$IntManifest.newArray(Manifest.scala:221)
	at scala.reflect.ManifestFactory$IntManifest.newArray(Manifest.scala:219)
	at scala.tools.nsc.ast.parser.ScannersCommon.createKeywordArray(Scanners.scala:72)
	at scala.tools.nsc.ast.parser.ScannersCommon.createKeywordArray$(Scanners.scala:68)
	at scala.tools.nsc.ast.parser.SyntaxAnalyzer.createKeywordArray(SyntaxAnalyzer.scala:20)
	at scala.tools.nsc.ast.parser.Scanners.$init$(Scanners.scala:1270)
	at scala.tools.nsc.ast.parser.SyntaxAnalyzer.<init>(SyntaxAnalyzer.scala:20)
	at scala.tools.nsc.Global$$anon$3.<init>(Global.scala:464)
	at scala.tools.nsc.Global.syntaxAnalyzer$lzycompute(Global.scala:462)
	at scala.tools.nsc.Global.syntaxAnalyzer(Global.scala:462)
	at scala.tools.nsc.Global.computeInternalPhases(Global.scala:665)
	at scala.tools.nsc.Global.computePhaseDescriptors(Global.scala:712)
	at scala.tools.nsc.Global.phaseDescriptors$lzycompute(Global.scala:719)
	at scala.tools.nsc.Global.phaseDescriptors(Global.scala:719)
	at scala.tools.nsc.Global$Run.<init>(Global.scala:1226)
	at scala.tools.nsc.Driver.doCompile(Driver.scala:47)
	at scala.tools.nsc.MainClass.doCompile(Main.scala:30)
	at scala.tools.nsc.Driver.process(Driver.scala:68)
	at scala.tools.nsc.Main.process(Main.scala)
	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
	at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
	at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
	at java.lang.reflect.Method.invoke(Method.java:498)
	at sbt.internal.inc.RawCompiler.getReporter$1(RawCompiler.scala:50)
	at sbt.internal.inc.RawCompiler.apply(RawCompiler.scala:71)
	at sbt.internal.inc.AnalyzingCompiler$.$anonfun$compileSources$6(AnalyzingCompiler.scala:350)
	at sbt.internal.inc.AnalyzingCompiler$$$Lambda$5500/1180376203.apply$mcV$sp(Unknown Source)
	at sbt.internal.inc.AnalyzingCompiler$.handleCompilationError$1(AnalyzingCompiler.scala:327)
	at sbt.internal.inc.AnalyzingCompiler$.$anonfun$compileSources$4(AnalyzingCompiler.scala:346)
	at sbt.internal.inc.AnalyzingCompiler$.$anonfun$compileSources$4$adapted(AnalyzingCompiler.scala:341)
	at sbt.internal.inc.AnalyzingCompiler$$$Lambda$5498/156164326.apply(Unknown Source)
	at sbt.io.IO$.withTemporaryDirectory(IO.scala:475)
[error] ## Exception when compiling 557 sources to /home/jenkins/workspace/scala-2.13.x-validate-main@3/build/quick/classes/library
```