@retronym 2019-04-24
@mkeskells @hrhino As discussed, here's an implementation of name table sharing across Global instances (an idea borrowed the idea from `dotc`.)

https://github.com/retronym/scala/pull/43

> the main motivation is that in our internal compilation we see that the name table is overloaded and has chains up to 500 long,

@hrhino Do you know what identifiers are leading this problem. I expect that #5474 would fix it, although it might also be possible to change the compiler or your plugin phases to avoid unnecessary freshening of some names.