@diesalbla 2020-05-08
@mkeskells Are you still working on this Pull Request? The results of those benchmarks looked promising.