@mkeskells 2020-05-09
> @mkeskells Are you still working on this Pull Request? The results of those benchmarks looked promising.

I would like to get back to it sometime, but working on 2.12 collection performance ATM, and likely for the next month at least. @diesalbla would you like to work on this together, or take this over?
@retronym has made some changes in the interim to make these methods less hot, so need to check to see the benefit