@dwijnand 2020-10-08
> @retronym has made some changes in the interim to make these methods less hot, so need to check to see the benefit, and the results that I quoted may well not still apply

Given that, and the time passed, shall we close this PR for now?  We can always reopen it if activity resumes.